Fix alignment of values within transaction details dialog
I was unable to align the link buttons with the other labels, so I changed the link buttons to labels and added a seperate link button for loading the block explorer next to it. This has the benefit of being able to select (highlight) the values in order to copy it.